+++ scripts/states/loading.js
@@ -1,0 +1,40 @@
+var Loading = function(game) {
+  var loading = {
+    init: function() {
+      game.add.sprite(0, 0, 'loading');
+    },
+
+    preload: function() {
+      // images
+      game.load.image('title', 'images/title.gif');
+      game.load.image('clear', 'images/clear.png');
+      game.load.image('white', 'images/white.png');
+      game.load.image('pink', 'images/pink.png');
+      game.load.image('yellow', 'images/yellow.png');
+      game.load.image('blue', 'images/blue.png');
+      game.load.image('orange', 'images/orange.png');
+      game.load.image('green', 'images/green.png');
+      game.load.image('gray', 'images/gray.png');
+      game.load.image('brown', 'images/brown.png');
+      game.load.image('waterfall', 'images/waterfall.gif');
+      game.load.image('hangar', 'images/hangar.gif');
+      game.load.spritesheet('victoryMsg', 'images/victoryMsg.png', 47, 6);
+
+      // sound
+      game.sfx = require('../sfx.js');
+      game.bgm = require('../music')();
+    },
+
+    create: function() {
+      game.input.gamepad.start();
+
+      game.state.add('splash', require('./splash.js')(game));
+      game.state.add('play', require('./play.js')(game));
+      game.state.start('splash');
+    }
+  };
+  
+  return loading;
+};
+
+module.exports = Loading;
