+++ scripts/states/splash.js
@@ -1,28 +1,5 @@
 var Splash = function(game) {
   var splash = {
-    init: function() {
-      // TODO: intro animation
-      game.sfx = require('../sfx.js');
-      game.bgm = require('../music')();
-    },
-
-    preload: function() {
-      // images
-      game.load.image('title', 'images/title.gif');
-      game.load.image('clear', 'images/clear.png');
-      game.load.image('white', 'images/white.png');
-      game.load.image('pink', 'images/pink.png');
-      game.load.image('yellow', 'images/yellow.png');
-      game.load.image('blue', 'images/blue.png');
-      game.load.image('orange', 'images/orange.png');
-      game.load.image('green', 'images/green.png');
-      game.load.image('gray', 'images/gray.png');
-      game.load.image('brown', 'images/brown.png');
-      game.load.image('waterfall', 'images/waterfall.gif');
-      game.load.image('hangar', 'images/hangar.gif');
-      game.load.spritesheet('victoryMsg', 'images/victoryMsg.png', 47, 6);
-    },
-
     create: function() {
       game.bgm.play('title.xm');
       game.add.sprite(0, 0, 'hangar');
